いいね機能の修正

+++ app/controllers/boil_posts_controller.rb
@@ -45,16 +45,20 @@ class BoilPostsController < ApplicationController
   end
 
   def like_create
-    @like = current_user.likes.new(boil_post_id: params[:id])
-    if @like.save!
-      redirect_to @boil_post
+    if @boil_post.user_id != current_user.id
+      @like = current_user.likes.new(boil_post_id: params[:id])
+      if @like.save!
+        redirect_to @boil_post
+      end
     end
   end
 
   def like_destroy
-    @like = Like.find_by(boil_post_id: params[:id]) 
-    if @like.destroy
-      redirect_to @boil_post
+    if @boil_post.user_id != current_user.id
+      @like = Like.find_by(boil_post_id: params[:id]) 
+      if @like.destroy
+        redirect_to @boil_post
+      end
     end
   end
